+++ lib/compass/commands/list_frameworks.rb
@@ -2,6 +2,7 @@ module Compass
   module Commands
     class ListFrameworks < ProjectBase
       attr_accessor :options
+      register :frameworks
       def initialize(working_path, options)
         super
       end
@@ -9,6 +10,11 @@ module Compass
       def execute
         Compass::Frameworks::ALL.each do |framework|
           puts framework.name unless framework.name =~ /^_/
+        end
+      end
+      class << self
+        def description(command)
+          "List the available frameworks"
         end
       end
     end